requirements.txt
azure-ai-generative==1.0.0b1
azure-identity==1.15.0
azure-ai-textanalytics==5.3.0
azure-search-documents==11.6.0b4
openai==1.47.0
requests==2.32.3
python-dotenv==1.0.1
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.2
jira==3.5.2
